requirements.txt
fastapi==0.79.0
uvicorn[standard]==0.18.2
pydantic==1.10.8
google-cloud-tasks==2.13.1
argparse==1.4.0